+++ plone/app/z3cform/interfaces.py
@@ -1,15 +1,29 @@
 # -*- coding: utf-8 -*-
 from plone.app.textfield.widget import IRichTextWidget as patextfield_IRichTextWidget  # noqa
-from z3c.form.interfaces import ISelectWidget as IBaseSelectWidget
 from z3c.form.interfaces import IFormLayer
+from z3c.form.interfaces import ISelectWidget as IBaseSelectWidget
 from z3c.form.interfaces import ISingleCheckBoxWidget
 from z3c.form.interfaces import ITextWidget
+from zope.interface import Interface
 from zope.schema.interfaces import IDate
 from zope.schema.interfaces import IDatetime
 
 
 class IPloneFormLayer(IFormLayer):
     """Request layer installed via browserlayer.xml"""
+
+
+class IFieldPermissionChecker(Interface):
+    """Adapter factory for checking whether a user has permission to
+    edit a specific field on a content object.
+    """
+
+    def validate(field_name, vocabulary_name=None):
+        """Returns True if the current user has permission to edit the
+        `field_name` field.  Returns False if the user does not have
+        permission.  Raises and AttributeError if the field cannot be
+        found.
+        """
 
 
 class IDateField(IDate):